Changed File structure to make it more organized
